[dram-core-prefetcher][tests] Fold trace-replay test into dram_core_large
Move test_dram_core_prefetcher_trace_replay out of its standalone file and
into test_prefetcher_BH_dram_core_large.py as a third scope, dropping the
duplicate boilerplate (pytestmark, _require_dram_core_prefetcher fixture,
shared imports) and the cross-file import of _bank_receivers_row_major.

Align the test with the parent branch's conventions while moving it:
- _require_dram_core_prefetcher autouse fixture instead of a raw env-var skip
- ttnn.init_device_compute_kernel_config(device.arch(), ...) instead of
  direct ttnn.WormholeComputeKernelConfig construction
- dram_core_prefetcher_session(device) context manager instead of a
  hand-rolled start/try/finally/stop
- math.ceil(K / ring_size) for K_per_shard, matching the file's other tests
